HOMEWORK_2/HOMEWORK_1/requirements.txt
faicons
markdown
openai
pyyaml
shiny
shinywidgets
plotly
pandas
ridgeplot
python-docx
requests
python-dotenv
jinja2